Use icecube.phys_services.ExtrudedPolygon instead of scipy.spatial.ConvexHull
